parking_system/requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0